fix(fsutil): fsync parent directory after atomic rename for crash safety
The atomic-write helpers used temp-file -> write -> fsync(file) -> rename but
never fsynced the parent directory. On Linux/Unix the rename is not durable
until the directory entry is flushed: after power loss the target could revert
to old content, or for a first-time write disappear, even though the write
"succeeded". The agent is exactly where this matters — a host losing power right
after a cert rotation could come back with a new host.crt but an old config.yml.

Extract the pattern into internal/fsutil.AtomicWriteFile (one helper instead of
three copies) and fsync the parent directory after the rename. The directory
sync is best-effort: unsupported on some filesystems, and a failure after a
successful rename is misleading, so it is logged rather than returned.

Call sites converted: the agent's host.crt / ca.crt / config.yml writes
(poller.go, was atomicWriteFile), SaveAgentConfig, and SaveServerConfig. The
old per-file helper and its tests move into the fsutil package.

Closes #225
